requirements.txt
ccxt
pandas
numpy
backtrader
python-binance
yfinance

alpaca-py
psycopg2-binary